requirements.txt
Flask==3.0.3
Flask-RESTful==0.3.9
SQLAlchemy==1.4.46
python-dotenv==0.21.0
Werkzeug==2.0.3
Flask-SQLAlchemy==2.5.1
psycopg2-binary==2.9.3
marshmallow
marshmallow-sqlalchemy
flask-restx
flask-cors
